requirements.txt
pybullet-svl==3.1.6.4
numpy==1.23.5
scipy==1.8.0
urdf-parser-py==0.0.3
pandas==1.4.2
ipython
plyfile

argparse
PyYAML
matplotlib
tqdm

pytest
networkx
seaborn
sklearn==0.0
scikit-learn==1.0.2
icecream
igibson==2.0.2
tensorboardX
Shapely~=1.8.2
open3d~=0.15.2

git+https://github.com/facebookresearch/pytorch3d.git@stable#=egg=pytorch3d
